move cache rh to another file, fix typo in readme - add _get_yt_proxy to BgUtilBaseGetPOTRH
